backend/requirements.txt
flask
flask-cors
flask-socketio
mysql-connector-python
waitress
flask-caching
